Update PS3.Rmd
Added extra commenting on the PlayGame function for better clarification of what the function is supposed to be doing.
Deleted the duplicative test on Door at line 98-100.
Added more commenting on the PlayGame test, with an example of what is supposed to happen.
Corrected the object name in the PlayGame test.