Update PS3.Rmd
Changed the object name in question 1 to be ContestantDoor, to save from confusion with the function Door being created in question 2.
Deleted the first Door Function, as it was unnecessary.
Deleted the comments informing how Door Function 2 was different than Door Function 1.
Clarified some wording in the Door function.
Altered the validate_door_integer function from checking if the object is an integer type to whether it is a whole number in the correct range.