Updated PS4 and wordcloud
Updated the word-cloud to verify the prism of colors corresponds to each word.
Removed the word-cloud color object.
Added a title and note to the word-cloud pdf to inform the viewer of what they are looking at.
Correct the Document Term Matrix code, at the end of the file. I turned the DTM into a table using tidy(), which then allowed data manipulation to complete the task.
Removed eval=FALSE on line 155, and then knitted the file.